Hw05 avatars

+++ controllers/contacts.js
@@ -5,9 +5,9 @@ const getAll = async (req, res, next) => {
   const { page = 1, limit = 20, favorite } = req.query;
   const skip = (page - 1) * limit;
   const { _id: owner } = req.user;
-  let query = { owner };
+  const query = { owner };
   if (favorite) {
-    query = { owner, favorite: true };
+    query.favorite = true;
   }
   const result = await Contacts.find(query, "", { skip, limit });
 
@@ -16,7 +16,10 @@ const getAll = async (req, res, next) => {
 
 const getById = async (req, res, next) => {
   const { contactId } = req.params;
-  const result = await Contacts.findById(contactId);
+  const { _id: owner } = req.user;
+  const result = await Contacts.findById(contactId)
+    .where("owner")
+    .equals(owner);
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -31,7 +34,13 @@ const add = async (req, res, next) => {
 
 const deleteById = async (req, res, next) => {
   const { contactId } = req.params;
-  console.log(contactId);
+  const { _id: owner } = req.user;
+  const contact = await Contacts.findById(contactId)
+    .where("owner")
+    .equals(owner);
+  if (!contact) {
+    throw HttpError(404, "Not found");
+  }
   const result = await Contacts.findByIdAndDelete(contactId);
   if (!result) {
     throw HttpError(404, "Not found");
@@ -41,6 +50,13 @@ const deleteById = async (req, res, next) => {
 
 const updateById = async (req, res, next) => {
   const { contactId } = req.params;
+  const { _id: owner } = req.user;
+  const contact = await Contacts.findById(contactId)
+    .where("owner")
+    .equals(owner);
+  if (!contact) {
+    throw HttpError(404, "Not found");
+  }
   const result = await Contacts.findByIdAndUpdate(contactId, req.body, {
     new: true,
   });
@@ -52,6 +68,13 @@ const updateById = async (req, res, next) => {
 
 const updateStatusContact = async (req, res, next) => {
   const { contactId } = req.params;
+  const { _id: owner } = req.user;
+  const contact = await Contacts.findById(contactId)
+    .where("owner")
+    .equals(owner);
+  if (!contact) {
+    throw HttpError(404, "Not found");
+  }
   const result = await Contacts.findByIdAndUpdate(contactId, req.body, {
     new: true,
   });
